Moved old Wifi code under next generation.
The next generation specification isn't yet followed, but rather the old code
is moved pretty much as it was expect renaming of some functions.
