+++ thali/NextGeneration/thaliWifiInfrastructure.js
@@ -1,6 +1,13 @@
 'use strict';
 
+var EventEmitter = require('events').EventEmitter;
+var inherits = require('util').inherits;
 var Promise = require('lie');
+var nodessdp = require('node-ssdp');
+var ip = require('ip');
+var crypto = require('crypto');
+
+var THALI_USN = 'urn:schemas-upnp-org:service:Thali';
 
 /** @module ThaliWifiInfrastructure */
 
@@ -39,9 +46,86 @@ var Promise = require('lie');
  * @fires event:networkChangedWifi
  * @fires discoveryAdvertisingStateUpdateWifiEvent
  */
-function ThaliWifiInfrastructure() {
-
+function ThaliWifiInfrastructure (deviceName, port) {
+  EventEmitter.call(this);
+  this.thaliUsn = THALI_USN;
+  this.deviceName = deviceName || crypto.randomBytes(16).toString('base64');
+  this.port = port || 0;
+  this.listening = null;
+  this.advertising = null;
+  // A variable to hold information about known peer availability states
+  // and used to avoid emitting peer availability changes in case the
+  // availability hasn't changed from the previous known value.
+  this.peerAvailabilities = {};
+  this._init(deviceName);
 }
+
+inherits(ThaliWifiInfrastructure, EventEmitter);
+
+ThaliWifiInfrastructure.prototype._init = function () {
+  var serverOptions = {
+    adInterval: 500,
+    allowWildcards: true,
+    logJSON: false,
+    logLevel: 'trace',
+    udn: this.deviceName
+  };
+  this._server = new nodessdp.Server(serverOptions);
+  this._setLocation();
+
+  this._client = new nodessdp.Client({
+    allowWildcards: true,
+    logJSON: false,
+    logLevel: 'trace'
+  });
+
+  this._client.on('advertise-alive', function (data) {
+    this._handleMessage(data, true);
+  }.bind(this));
+
+  this._client.on('advertise-bye', function (data) {
+    this._handleMessage(data, false);
+  }.bind(this));
+};
+
+ThaliWifiInfrastructure.prototype._setLocation = function (address, port, path) {
+  address = address || ip.address();
+  port = port || this.port;
+  path = path || 'NotificationBeacons';
+  this._server._location = 'http://' + address + ':' + port + '/' + path;
+};
+
+ThaliWifiInfrastructure.prototype._handleMessage = function (data, available) {
+  if (this._shouldBeIgnored(data)) {
+    return;
+  }
+  var peer = {
+    peerIdentifier: data.USN,
+    peerLocation: data.LOCATION,
+    peerAvailable: available
+  };
+  if (this.peerAvailabilities[peer.peerIdentifier] === available) {
+    return;
+  }
+  this.peerAvailabilities[peer.peerIdentifier] = available;
+  this.emit('wifiPeerAvailabilityChanged', [peer]);
+};
+
+// Function used to filter out SSDP messages that are not
+// relevant for Thali.
+ThaliWifiInfrastructure.prototype._shouldBeIgnored = function (data) {
+  // First check if the data contains the Thali-specific USN.
+  if (data.USN.indexOf(this.thaliUsn) >= 0) {
+    // We also discover ourself via SSDP to need to filter
+    // out the messages that are originating from this device.
+    if (data.USN.indexOf(this.deviceName) === 0) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+  return true;
+};
 
 /**
  * This method MUST be called before any other method here other than
@@ -101,10 +185,18 @@ ThaliWifiInfrastructure.prototype.stop = function () {
  *
  * @returns {Promise<?Error>}
  */
-ThaliWifiInfrastructure.prototype.startListeningForAdvertisements =
-  function () {
-    return new Promise();
-  };
+ThaliWifiInfrastructure.prototype.startListeningForAdvertisements = function () {
+  var self = this;
+  if (this.listening) {
+    return Promise.resolve();
+  }
+  this.listening = true;
+  return new Promise(function(resolve, reject) {
+    self._client.start(function () {
+      resolve();
+    });
+  });
+};
 
 /**
  * This will stop the local Wi-Fi Infrastructure Mode discovery mechanism
@@ -124,7 +216,16 @@ ThaliWifiInfrastructure.prototype.startListeningForAdvertisements =
  * @returns {Promise<?Error>}
  */
 ThaliWifiInfrastructure.prototype.stopListeningForAdvertisements = function () {
-  return new Promise();
+  var self = this;
+  if (!this.listening) {
+    return Promise.resolve();
+  }
+  this.listening = false;
+  return new Promise(function(resolve, reject) {
+    self._client.stop(function () {
+      resolve();
+    });
+  });
 };
 
 /**
@@ -181,7 +282,22 @@ ThaliWifiInfrastructure.prototype.stopListeningForAdvertisements = function () {
  * @returns {Promise<?Error>}
  */
 ThaliWifiInfrastructure.prototype.startUpdateAdvertisingAndListening = function() {
-  return new Promise();
+  var self = this;
+  // TODO: USN should be regenerated every time this method is called, because
+  // according to the specification, that happens when the beacon string is changed.
+  // Is below enough or should we use some uuid library or something else?
+  var randomString = crypto.randomBytes(16).toString('base64');
+  // TODO: Appends to USN list, but does not remove.
+  this._server.addUSN(this.thaliUsn + '::' + randomString);
+  if (this.advertising) {
+    return Promise.resolve();
+  }
+  this.advertising = true;
+  return new Promise(function(resolve, reject) {
+    self._server.start(function () {
+      resolve();
+    });
+  });
 };
 
 /**
@@ -199,7 +315,16 @@ ThaliWifiInfrastructure.prototype.startUpdateAdvertisingAndListening = function(
  * @returns {Promise<?Error>}
  */
 ThaliWifiInfrastructure.prototype.stopAdvertisingAndListening = function() {
-  return new Promise();
+  var self = this;
+  if (!this.advertising) {
+    return Promise.resolve();
+  }
+  this.advertising = false;
+  return new Promise(function(resolve, reject) {
+    self._server.stop(function () {
+      resolve();
+    });
+  });
 };
 
 /**
